Change maven project version to 1.5.9-SNAPSHOT
The "equals()" implementation of com.bc.fiduceo.geometry.Point and subclasses has been changed to override Object.equals().
This change allows this equals method to be used in JUnit's assertEquals as usual.

SmapReader.extractGeoMinMax() has been changed to deal with the fact that the products are verically flipped and latitude values are flipped as well.